--- a/apps/iconlaunch/metadata.json
+++ b/apps/iconlaunch/metadata.json
@@ -10,7 +10,8 @@
 	"supports": ["BANGLEJS2"],
 	"storage": [
 		{ "name": "iconlaunch.app.js", "url": "app.js" },
-		{ "name": "iconlaunch.settings.js", "url": "settings.js" }
+		{ "name": "iconlaunch.settings.js", "url": "settings.js" },
+		{ "name": "iconlaunch.na.img", "url": "na.img" }
 	],
 	"data": [{"name":"iconlaunch.json"},{"name":"iconlaunch.cache.json"}],
 	"screenshots": [{ "url": "screenshot1.png" }, { "url": "screenshot2.png" }],