Update:
Firmware change in 2v185 got `drawImage` much faster so baseline is now 23ms per `drawItem` and 420ms startup time with populated cache.
* Reduces array access via index to prevent list searching
* Only clear item area if needed
* Real icon for apps not having an icon (no performance benefit, but better icon :)
* Presort apps into scroller items for faster access
* jit for `drawItem`

Overall result is 460ms startup time and 19ms per `drawItem`, so ~9% longer startup for ~20% faster drawing during scrolling.