@gfwilliams 2023-07-24
That's great - thanks! You happy to merge?

I have to admit I'm very surprised about drawImages being faster though! it scans through every image for every pixel rendered - but I guess maybe because it's only rendering a slice of an image that is actually more efficient