@gfwilliams 2023-07-24
It'd actually be interesting to do some benchmarks on drawImage - it looks like maybe by default we iterate over all pixels in the image regardless of how many are on screen - so it might be possible to improve the logic to at the very least skip the top + bottom rows that aren't needed