@halemmerich 2023-07-24
This works fine for me but currently the precaching of icons trades some better smoothness during the initial scroll for a bit of additional startup time for the caching. I'm not yet sure if I like this tradeoff and might go back on that. Needs a bit more benchmarking to decide this.
The differences between `g.drawImage` and `g.drawImages` are certainly interesting. I had just blamed that on giving the whole work in form of the layers object to native code at once instead of doing it once at a time in a JS-loop. We might have some performance to gain here for all apps :)
At least for scrollers usually the top and bottom items together can be redrawn using the pixel count of only one item since both are cut off, effectively removing one item worth of pixels to draw.