@halemmerich 2023-07-24
> it looks like maybe by default we iterate over all pixels in the image

It sure does:
```
drawImage                   21.59729003906
drawImages overlapping      41.20292663574
drawImages not overlapping  41.26739501953
drawImages half optioned    22.51243591308
drawImages full optioned    0.89416503906
drawImage half clipped      21.39587402343
drawImages half clipped     22.62573242187
drawImage full clipped      22.45178222656
drawImages full clipped     0.75035095214
```
Clipping via `setClipRect` and otherwise via `drawImages` options object. Times are averaged over 400 images drawn. `drawImages` used a layers object with 4 images.

So in conclusion `drawImages` is indeed slower as you said until enough pixels are clipped of to make up for the overhead of scanning all images. The additional improvement bringing it over the edge in this case probably came from restructuring the loop code to build the layer object.