@gfwilliams 2023-07-25
Thanks for those benchmarks!

> Why does drawImages scan through all images

drawImages scans over the area, checking every image for each pixel and drawing the pixel *once* only.

It's done this way specifically to remove any chance of flicker on Bangle.js 1. Obviously if you can find a way to optimise it, PRs welcome, but the current solution is reasonably tidy.

Please can you try the latest cutting edge build now with a drawImage benchmark? It should be a lot faster - I think with this, you could probably revert the `drawImage->drawImages` changes and you'd see a decent speed boost?