@halemmerich 2023-07-25
The benchmark shows good improvements:
```
drawImage                   21.60003662109
drawImages overlapping      41.23794555664
drawImages not overlapping  41.23626708984
drawImages half optioned    22.52113342285
drawImages full optioned    0.82275390625
drawImage half clipped      12.56256103515
drawImages half clipped     22.62023925781
drawImage full clipped      3.20739746093
drawImages full clipped     0.67245483398
```

For `drawItem` I now get about 2ms less than the last end result, so it should be 8ms all in all since the original step from `drawImage` to `drawImages` saved 6ms. So pretty great result overall I would say.