@KTibow 2023-08-31
This PR might've broken some stuff, I can't elaborate more because I'm on mobile but before I updated the launcher worked, and now other icons get randomly hidden when I select an icon