@halemmerich 2023-08-31
Do you use the direct launch setting? If not, it might be a problem that does no longer show up on the current nightly firmwares.