@KTibow 2023-08-31
Okay, I'm fine waiting for the firmware to update then.
Question: I don't have enough icons to scroll. Do I need to worry about the launcher's performance going down because of this?