@halemmerich 2023-09-01
If you don't have enough icons to scroll, scrolling will only redraw what is already there in the worst case. In that case it is actually pretty irrelevant how long the redraw takes :)
As for the different icon, yes that would be possible. Showing the first letter is a good idea for this. What app is shown that does not have an icon for you? Probably that app should just get an icon.