@KTibow 2023-09-01
> If you don't have enough icons to scroll, scrolling will only redraw what is already there in the worst case

I was more talking about startup performance - but it probably doesn't matter that much given I can't tell if it's slower.

> What app is shown that does not have an icon for you? Probably that app should just get an icon.

Just some custom apps I coded.